[Refactor] cardSlice 이름 변경 -> informationSlice
