+++ src/pages/event/ui/Event.jsx
@@ -1,12 +1,15 @@
-import {useEffect, useState} from "react";
+import {useEffect} from "react";
 import {useDispatch} from "react-redux";
+import {useNavigate} from "react-router-dom";
 import Common from "@pages/common/event&policyRec/ui/Common.jsx";
 import {dummyData} from "@test/eventDummy.js";
-import {add} from "@stores/slices/cardSlice.js";
+import {add} from "@stores/slices/informationSlice.js";
 import {setPage} from "@stores/slices/pagingSlice.js";
+
 
 const Event = () => {
   const dispatch = useDispatch();
+  const navigate = useNavigate();
 
   useEffect(() => {
     dummyData.forEach((item) => {
@@ -16,8 +19,12 @@ const Event = () => {
     dispatch(setPage({itemsCount: 22, totalItems: 160, pageRange:5}))
   }, []);
 
+  const onClick= (id) => {
+
+  }
+
   return (
-      <Common />
+      <Common onClick={onClick}/>
   );
 };
