+++ src/stores/slices/index.js
@@ -1,2 +1,2 @@
-export {default as cardSlice} from "./cardSlice.js";
+export {default as informationSlice} from "./informationSlice.js";
 export {default as pagingSlice} from "./pagingSlice.js";
